Chore: Prepare batching of updates/creations: partitionWith
In order to batch-process we need to distinguish between deals with an
existing page which we simply update and deals that need to have their
page created.  Currently we do one pass through a list of mixed deals.
This introduces `partitionWith` in order to partition the `allDeals`
list and also retrieve their dealId.

The resulting lists can then be handled separately in batches.